License change
Okay, so after being called out and researching a bit, it turns out it's
not possible under US law to publish works directly to the public
domain. I did a bit of digging and decided to license the library under
the MIT license instead, as it seems to be the most permissive and
easily understood.