Introduction of password strength check functionality.
This commit introduces a new namespace intended to facilitate the
addition of new functionaility for calculating password strength based
on various configurable algorithms. Initially, the only available
algorithm is an adaptation of the Wolfram|Alpha password strength
calculator. The main differences are the lack of Wolfram's "extra
criteria" and dictionary lookup.

The new class can be used standalone to generate a numeric score.

    <?php
    $calculator = new Phpass\Strength\Wolfram;
    $strength = $calculator->check('MySuperSecretPassword');

This functionality is considered alpha phase, and will not be integrated
into the PEAR release or the main docs yet.